+++ CHANGELOG.md
@@ -6,6 +6,7 @@ please submit a pull request fixing it.
 
 ## Unreleased - 0.8.6
 - Fixed bug in rendering of registration form. ([#222])
+- User profile OGP tags ([#228])
 
 ## 0.8.5 - December 31st, 2021
 - Changes to the config file:
@@ -175,3 +176,4 @@ please submit a pull request fixing it.
 [#218]: https://github.com/rcos/Telescope/pull/218
 [#220]: https://github.com/rcos/Telescope/pull/220
 [#222]: https://github.com/rcos/Telescope/pull/222
+[#228]: https://github.com/rcos/Telescope/pull/228