+++ CHANGELOG.md
@@ -7,6 +7,7 @@ please submit a pull request fixing it.
 ## Unreleased - 0.8.6
 - Fixed bug in rendering of registration form. ([#222])
 - Added meeting counts and filtering by semester in admin panel. 
+- User profile OGP tags ([#228])
 
 ## 0.8.5 - December 31st, 2021
 - Changes to the config file:
@@ -176,3 +177,4 @@ please submit a pull request fixing it.
 [#218]: https://github.com/rcos/Telescope/pull/218
 [#220]: https://github.com/rcos/Telescope/pull/220
 [#222]: https://github.com/rcos/Telescope/pull/222
+[#228]: https://github.com/rcos/Telescope/pull/228